회원가입/로그인 구현

+++ src/routes/posts.js
@@ -1,14 +1,17 @@
 const { Router } = require('express');
-const ApiError = require('../utils/error');
-const { Post } = require('../db/models');
+const jwt = require('jsonwebtoken');
 const { BaseError } = require('sequelize');
+
+const ApiError = require('../utils/error');
+const { Post, User } = require('../db/models');
+const { isLoggedIn } = require('../middlewares/auth');
 
 const router = Router();
 
-router.post('/', async (req, res, next) => {
+router.post('/', isLoggedIn, async (req, res, next) => {
   try {
-    console.log(req.body);
     const { title, content } = req.body;
+    const { userId } = jwt.decode(req.cookies.accessToken);
 
     if (!title && !content)
       throw new ApiError(412, '데이터 형식이 올바르지 않습니다.');
@@ -19,26 +22,38 @@ router.post('/', async (req, res, next) => {
         throw new ApiError(412, '게시글 내용의 형식이 일치하지 않습니다.');
     }
 
-    await Post.create({ title, content });
+    const user = await User.findByPk(userId);
+
+    await Post.create({ title, content, userId: user.userId });
 
     res.status(201).json({ message: '게시글 작성에 성공하였습니다.' });
   } catch (err) {
     if (err instanceof ApiError) next(err);
-    else next(new ApiError(400, '게시글 작성에 실패하였습니다.'));
+    else next(new ApiError(400, '게시글 작성에 실패하였습니다.', err.stack));
   }
 });
 
 router.get('/', async (req, res, next) => {
   try {
     const posts = await Post.findAll({
-      attributes: ['postId', 'title', 'createdAt', 'updatedAt'],
+      attributes: ['postId', 'userId', 'title', 'createdAt', 'updatedAt'],
+      include: {
+        model: User,
+        attributes: ['nickname'],
+      },
       order: [['postId', 'DESC']],
+      raw: true,
+    });
+
+    posts.forEach((post) => {
+      post.nickname = post['User.nickname'];
+      delete post['User.nickname'];
     });
 
     res.status(200).json({ data: posts });
   } catch (err) {
     if (err instanceof ApiError) next(err);
-    else next(new ApiError(400, '게시글 조회에 실패하였습니다.'));
+    else next(new ApiError(400, '게시글 조회에 실패하였습니다.', err.stack));
   }
 });
 
@@ -46,19 +61,29 @@ router.get('/:postId', async (req, res, next) => {
   try {
     const { postId } = req.params;
 
-    const post = await Post.findByPk(postId);
+    const post = await Post.findByPk(postId, {
+      include: {
+        model: User,
+        attributes: ['nickname'],
+      },
+      raw: true,
+    });
+
+    post.nickname = post['User.nickname'];
+    delete post['User.nickname'];
 
     res.status(200).json({ data: post });
   } catch (err) {
     if (err instanceof ApiError) next(err);
-    else next(new ApiError(400, '게시글 조회에 실패하였습니다.'));
+    else next(new ApiError(400, '게시글 조회에 실패하였습니다.', err.stack));
   }
 });
 
-router.put('/:postId', async (req, res, next) => {
+router.put('/:postId', isLoggedIn, async (req, res, next) => {
   try {
     const { title, content } = req.body;
     const { postId } = req.params;
+    const { userId } = jwt.decode(req.cookies.accessToken);
 
     if ((!title && !content) || isNaN(postId))
       throw new ApiError(412, '데이터 형식이 올바르지 않습니다.');
@@ -70,6 +95,9 @@ router.put('/:postId', async (req, res, next) => {
     }
 
     const post = await Post.findByPk(postId);
+    const user = await User.findByPk(userId);
+
+    if (post.userId !== user.userId) throw new Error();
 
     await post.update({ title, content });
 
@@ -77,18 +105,23 @@ router.put('/:postId', async (req, res, next) => {
   } catch (err) {
     if (err instanceof ApiError) next(err);
     else if (err instanceof BaseError)
-      next(new ApiError(401, '게시글이 정상적으로 수정되지 않았습니다.'));
-    else next(new ApiError(400, '게시글 수정에 실패하였습니다.'));
+      next(
+        new ApiError(401, '게시글이 정상적으로 수정되지 않았습니다.', err.stack)
+      );
+    else next(new ApiError(400, '게시글 수정에 실패하였습니다.', err.stack));
   }
 });
 
-router.delete('/:postId', async (req, res, next) => {
+router.delete('/:postId', isLoggedIn, async (req, res, next) => {
   try {
     const { postId } = req.params;
+    const { userId } = jwt.decode(req.cookies.accessToken);
 
     const post = await Post.findByPk(postId);
+    const user = await User.findByPk(userId);
 
     if (!post) throw new ApiError(404, '게시글이 존재하지 않습니다.');
+    if (post.userId !== user.userId) throw new Error();
 
     await Post.destroy({ where: { postId } });
 
@@ -97,7 +130,7 @@ router.delete('/:postId', async (req, res, next) => {
     if (err instanceof ApiError) next(err);
     else if (err instanceof BaseError)
       next(new ApiError(401, '게시글이 정상적으로 삭제되지 않았습니다.'));
-    else next(new ApiError(400, '게시글 작성에 실패하였습니다.'));
+    else next(new ApiError(400, '게시글 작성에 실패하였습니다.', err.stack));
   }
 });
